compositor: start the dbus session in the compositor
We were starting the dbus session in the orbital-client script wrapper.
That approach was not working properly, since every time orbital-client
was restarted a new session was started. Additionally, the compositor
didn't know of it, so clients autostarted did not have access to it.
Instead, let the compositor launch dbus-launch and set the
appropriate environment variables.