both: properly handle the lock screen when switching vt
This patch ensures that we have a lock screen visible before doing
a vt switch. That means that we won't see a glimpse of the screen
when going back to the vt after the session assigned to it was
locked.
To enable that this patch adds a fake repaint loop running at 10Hz
which runs when the session is inactive. This allows the shell client
to still handle DBus signals and send Wayland requests.