both: fix orbital-client crashes when adding an output
On output creation the compositor was trying to send the
desktop_shell_load_output event with a resource that was often null
since the client didn't have the time to bind the wl_output yet.
Instead wait for the client to signal the compositor it knows the
output before sending the load event. The downside is that we now
incur a roundtrip, but it will hardly be noticeable, since bringing
up an output takes order of magnitudes more time than a roundtrip.