compositor: modify the Xwayland internal position of X windows
To make multiple views for an X window work correctly wrt. input, make the
surface with the pointer focus be always at 0x0, and the others at 10000x10000.
The Xwayland internal representation of windows doesn't really need to follow
reality, the important thing is that the window that is supposed to get input
events is inside the screen rect.