all: introduce a global interface to authorize binding restricted interfaces
The new orbital_authorizer global interface allows a client to ask
for the authorization to bind to restricted interfaces such as
orbital_screenshooter.
Currently the authorization process is hardcoded and only the
orbital-screenshooter program is allowed to bind orbital_screenshooter,
anything else will fail. The goal is to determine whether a client
is allowed or not by reading some configuration file.